Move NonDaemonThreadLogger to package cats.effect
