+++ core/jvm/src/main/scala/cats/effect/IOApp.scala
@@ -17,7 +17,6 @@
 package cats.effect
 
 import scala.concurrent.{blocking, CancellationException}
-import internals.NonDaemonThreadLogger
 
 import java.util.concurrent.CountDownLatch
